+++ src/atmosphere.h
@@ -1,5 +1,5 @@
 #ifndef _ATMOSPHERE_H_
 #define _ATMOSPHERE_H_
-double air_density(double R);
-double geopotential_height(double R);
+float air_density(float R);
+float geopotential_height(float R);
 #endif
